requirements.txt
alembic==1.5.2
appdirs==1.4.4
APScheduler==3.7.0
argh==0.26.2
bcrypt==3.2.0
black==20.8b1
blinker==1.4
boto3==1.17.107
botocore==1.20.107
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2021.5.30
cffi==1.14.4
charset-normalizer==2.0.3
click==7.1.2
colorhash==1.0.3
configparser==5.0.2
cryptography==3.3.2
cssselect2==0.4.1
dataclasses==0.8
defusedxml==0.7.1
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Alembic==2.0.1
Flask-API==2.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Images==3.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-MonitoringDashboard==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-User==1.0.2.2
Flask-WTF==0.14.3
idna==3.1
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.10.0
Mako==1.1.4
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numpy==1.19.5
pandas==1.1.5
passlib==1.7.4
pathspec==0.8.1
PIL==1.1.7
pilkit==2.0
Pillow==8.2.0
pillowcase==2.0.0
psutil==5.8.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
redis==3.5.3
regex==2020.11.13
requests==2.26.0
s3transfer==0.4.2
scipy==1.5.4
six==1.15.0
SQLAlchemy==1.3.22
tinycss2==1.1.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.6
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
